Theck for primary_group_id to be set rather primary_group.
This fixes a useless query if there's no primary_group_id set because django
tries to evaluate if `primary_group` is set and queries for it.  This
changeset does not fix the problem that `primary_group` is not eagerloadable
but reduces the amount of queries on articles with much comments quite a lot.