Found a way to make the topiclist efficient again.
This changeset introduces a way to split the view into two queries.  All selecting and sorting is done in the first query that only selects the topic ids we need to display.  The second query is only a range select on those ids to query all information we need to display the template properly.  The second query selects the same amount of rows every time so it's highly efficient.  The first though still depends on the page 50 limit we have as LIMIT on mysql isn't fast and will never be :( But it selects less data so that this query is much faster than before