Rewritten search result highlighting.  We now get better results most of the time.
There are though some glitchy parts resulting in fancy output but this comes from not stripped-down content from the search index.  We need to find a way to create an index not using the rendered html but using the raw-text.
It would be much better if we can parse the text in the index-process and strip-down needless macros and other parts.